Move `splitParent` from `string.ts` into tests folder
