+++ news/3 Code Health/3988.md
@@ -1,0 +1,1 @@
+Move `splitParent` from `string.ts` into tests folder.
